Split scenes
and codewars.js fix
